+++ crypto-mac/Cargo.toml
@@ -13,15 +13,16 @@ categories = ["cryptography", "no-std"]
 
 [dependencies]
 generic-array = "0.14"
+crypto-common = { version = "0.1", path = "../crypto-common/" }
 subtle = { version = "2", default-features = false }
 
 blobby = { version = "0.3", optional = true }
-cipher = { version = "=0.3.0-pre.4", optional = true, path = "../cipher" }
 rand_core = { version = "0.6", optional = true }
 
 [features]
 dev = ["blobby"]
-std = []
+core-api = ["crypto-common/core-api"]
+std = ["crypto-common/std"]
 
 [package.metadata.docs.rs]
 all-features = true
